corregir path / probando DAO posta

+++ salidas-data.php
@@ -11,15 +11,14 @@ switch ($_GET["act"])
     case 'edit':
 
 
-
         $territoriosDAO = new \DAO\CoreDAO($db,"salidas_territorios");
 
         $territorios=$_POST["salida_territorios"];
 
         unset($_POST["salida_territorios"]);
 
-
         $res=$DAO->upsert($_POST);
+
 
         if($res)
         {
@@ -29,30 +28,41 @@ switch ($_GET["act"])
             $salidasTerritorio=array();
             foreach ($territorios as $k=>$v)
             {
-                $t=array(
-                    "territorio"=>$v["numero"],
-                    "salida"=>$res
-                );
 
+                if(is_numeric($v["numero"]) || !$v["numero"])
+                {
 
-                if($v["id"])
-                {
-                    $t["salidas_territorios_id"]=$v["id"];
-                }
-
-                if(!$v["delete"])
-                {
-                    $t= $territoriosDAO->upsert($t);
-                }
-                else
-                {
-                    $t= $territoriosDAO->delete(
-                        array(
-                            "salidas_territorios_id"=>$v["id"]
-                        )
+                    $t=array(
+                        "territorio"=>$v["numero"],
+                        "salida"=>$res
                     );
 
+
+                    if($v["id"])
+                    {
+                        $t["salidas_territorios_id"]=$v["id"];
+                    }
+
+
+                    if(!$v["delete"])
+                    {
+                        $t= $territoriosDAO->upsert($t);
+
+
+                    }
+                    else
+                    {
+                        $t= $territoriosDAO->delete(
+                            array(
+                                "salidas_territorios_id"=>$v["id"]
+                            )
+                        );
+
+
+                    }
+
                 }
+
 
 
                // $salidasTerritorio[]= $t;
